Do the same special casing as last commit, but for `lines_lf`.
I also tested similar optimizations in the other modules, but it
doesn't give consistently better results in those cases.  In the
case of the utf16 module, it's because it already coincidentally
had a similar optimization due to how it needs to work for correct
results.  For lines_unicode and lines_crlf, the inherent higher
complexity of what they're doing on a per-byte basis seems to
thwart any meaningful benefits of special casing short strings.