Optimize lines_crlf
This change is not ready to be merged because it does not yet support x86. I
don't know how to do the operation we need with SSE intrinsics.

This new approach adds a another method to ByteChunk that lets you shift a value
across two vectors. This let's us turn a 2-pass counting algorithim into a
single pass. Making this change alone resulted in a 40% speedup.

There is also some loop unrolling, but there was not a large difference in
unrolling over a factor of 2.