Fix CI after uncontrollable dev-dependency update.
Criterion has a transitive dependency that we can't control the version
of, which requires > Rust 1.65.  This means we can no longer run tests
using that version, even though it's our MSRV.  So switch to just
building with our MSRV rather than building *and* testing.  In practice
this should be fine, as building should be enough since we're running
tests with latest stable anyway.